requirements.txt
selenium>=4.15.2
webdriver-manager>=4.0.1
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

